requirements.txt
bidict==0.23.1
blinker==1.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
opentelemetry-api==1.25.0
opentelemetry-distro==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
protobuf==4.25.3
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.32.3
setuptools==70.2.0
simple-websocket==1.0.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
